// Carbazaar_frontend/src/app/interfaces/oldcardetails.ts
export interface oldCarDetails {
  oldCarName: string;
  oldCarVariant: string;
  history: string;
  ownerType: string;
  kilometerDriven: any;
  fuelType: string;
  transmission: string;
  registration: string;
  insuranceLastDate: string;
  insuranceDescription: string;
  yearOfPurchase: string;
  cdPlayer: boolean;
  mp3Player: boolean;
  seatUpholseryDetails: string;
  frontTyreDesc: string;
  auxCompatibility: boolean;
  rearTyreDesc: string;
  usbCompatibility: boolean;
  speakers: any;
}